services/auth_service/requirements.txt
fastapi==0.120.2
pydantic[email]==2.12.3
uvicorn==0.38.0
pydantic-settings==2.11.0
SQLAlchemy==2.0.44
asyncpg==0.30.0
faststream[rabbit]==0.6.1
orjson==3.11.4
argon2-cffi==25.1.0
pyjwt[crypto]==2.10.1
